Fix xieite::system::StandardStreamsController::readString

+++ include/xieite/math/robert.hpp
@@ -7,7 +7,7 @@
 
 namespace xieite::math {
 	template<std::integral Integral>
-	inline constexpr Integral robert = std::numeric_limits<Integral>::max() / std::numbers::phi_v<long double>;
+	inline constexpr Integral robert = std::numeric_limits<Integral>::max() / std::numbers::phi_v<double>;
 }
 
 // Thanks to melak47 for naming this constant
